+++ src/graphql/elasticsearch/catalog/resolver.js
@@ -1,12 +1,19 @@
 import config from 'config';
 import client from '../client';
-import { buildQuery } from '../queryBuilderVSF';
+import { buildQuery } from '../queryBuilder';
 
-async function list(filter, sort, currentPage, pageSize, search,) {
+async function list(filter, sort, currentPage, pageSize, search, context, rootValue) {
   let query = buildQuery(filter, sort, currentPage, pageSize, search);
 
+  const parseURL = context.req.url.replace(/^\/+|\/+$/g, '');
+  let urlParts = parseURL.split('/');
+  if (urlParts.length < 1) {
+      throw new Error('Please provide following parameters: /graphql/<storeId>/');
+  }
+  const storeId = parseInt(urlParts[0])
+
   const response = await client.search({
-    index: config.elasticsearch.indices[0],
+    index: config.elasticsearch.indices[storeId],
     type: config.elasticsearch.indexTypes[0],
     body: query
   });
@@ -16,8 +23,8 @@ async function list(filter, sort, currentPage, pageSize, search,) {
 
 const resolver = {
   Query: {
-    products: (_, { search, filter, sort, currentPage, pageSize }) =>
-      list(filter, sort, currentPage, pageSize, search)
+    products: (_, { search, filter, sort, currentPage, pageSize }, context, rootValue) =>
+      list(filter, sort, currentPage, pageSize, search, context, rootValue)
   }
 };
